backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
supabase==2.4.5
httpx==0.27.2
jinja2==3.1.2
weasyprint==52.5
mysql-connector-python==8.2.0
cloudinary==1.34.0
